fixed folder structure, WIP Sign In and user profile
