+++ .env
@@ -1,3 +1,3 @@
-REACT_APP_BACKEND_URL=http://localhost:8080
+REACT_APP_BACKEND_URL=http://localhost:5000
 REACT_APP_FIREBASE_API_KEY=AIzaSyAPH8hqLKjFryeuuMx8vgYLUg1rXEvrlf8
 REACT_APP_FIREBASE_AUTH_DOMAIN=cmpt470-316711.firebaseapp.com
