+++ package.json
@@ -11,6 +11,7 @@
     "axios": "^0.21.1",
     "bootstrap": "^5.0.2",
     "firebase": "^8.8.1",
+    "jshint": "^2.13.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-firebaseui": "^5.0.2",
