+++ src/App.js
@@ -1,14 +1,12 @@
 import React from 'react';
+import { BrowserRouter as Router, Switch, Route} from 'react-router-dom';
 import './App.css';
 import Navbar from "./components/Navbar/Navbar"
-import HomePage from './HomePage'
-import Header from './components/topMenu'
-import HomeContent from './components/HomePageContent/HomePageContent'
-import ShowClientsAndPhotographers from './components/showClients/showClientsAndPhotographers';
-import AboutUs from './AboutUs';
-import Booking from './Booking';
-import { BrowserRouter as Router, Switch, Route} from 'react-router-dom';
-import SignInScreen from './Firebase'
+import HomePage from './components/HomePage/HomePage'
+import AboutUs from './components/AboutUs/AboutUs';
+import Booking from './components/Booking/Booking';
+import SignInScreen from './components/SignInUp/Firebase'
+import UserProfile from './components/UserProfile/UserProfile'
 
 
 function App() {
@@ -16,13 +14,14 @@ function App() {
     return (
         <Router>
         <Switch>
-            <div>
+            <>
                 <Navbar />
                 <Route exact path="/" component={HomePage} />
                 <Route path="/about" component={AboutUs} />
                 <Route path="/booking" component={Booking} />
                 <Route path="/login" component={SignInScreen} />
-            </div>
+                <Route path="/userProfile" component={UserProfile} />
+            </>
         </Switch>
         </Router>
   );
