+++ src/components/SignInUp/Firebase.js
@@ -1,5 +1,5 @@
 // import firebase from "firebase/app";
-// import "firebase/auth";
+// import firebase from "firebase/auth";
 import firebase from "firebase";
 import React from 'react';
 import StyledFirebaseAuth from 'react-firebaseui/StyledFirebaseAuth';
@@ -17,7 +17,7 @@ const uiConfig = {
 	// Popup signin flow rather than redirect flow.
 	signInFlow: 'popup',
 	// Redirect to /signedIn after sign in is successful. Alternatively you can provide a callbacks.signInSuccess function.
-	signInSuccessUrl: '/',
+	signInSuccessUrl: '/userProfile',
 	// We will display Google and Facebook as auth providers.
 	signInOptions: [
 			firebase.auth.GoogleAuthProvider.PROVIDER_ID,
@@ -29,9 +29,7 @@ const uiConfig = {
 function SignInScreen() {
 	return (
 		<div>
-		<h1>My App</h1>
-		<p>Please sign-in:</p>
-		<StyledFirebaseAuth uiConfig={uiConfig} firebaseAuth={firebase.auth()} />
+			<StyledFirebaseAuth uiConfig={uiConfig} firebaseAuth={firebase.auth()} />
 		</div>
 	);
 }
