+++ src/components/UserProfile/UserProfile.jsx
@@ -1,0 +1,9 @@
+function UserProfile() {
+	return (
+		<div>
+			UserProfile
+		</div>
+	);
+}
+
+export default UserProfile
